separete API call api from Login component
